labai_fastapi/requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
pypdf
python-docx
python-multipart
jinja2
python-dotenv
